+++ DatadogInternal/Sources/NetworkInstrumentation/URLSession/URLSessionTaskSwizzler.swift
@@ -44,7 +44,14 @@ internal final class URLSessionTaskSwizzler {
         private let method: Method
 
         static func build() throws -> TaskResume {
-            return try TaskResume(selector: self.selector, klass: URLSessionTask.self)
+            // RUM-2690: We swizzle private `__NSCFLocalSessionTask` class as it appears to be uniformly used
+            // in iOS versions 12.x - 17.x. Swizzling the public `URLSessionTask.resume()` doesn't work in 12.x and 13.x.
+            // See https://github.com/DataDog/dd-sdk-ios/pull/1637 for full `URLSessionTask` class dumps in major iOS versions.
+            let className = "__NSCFLocalSessionTask"
+            guard let klass = NSClassFromString(className) else {
+                throw InternalError(description: "Failed to swizzle `URLSessionTask`: `\(className)` class not found.")
+            }
+            return try TaskResume(selector: self.selector, klass: klass)
         }
 
         private init(selector: Selector, klass: AnyClass) throws {